requirements.txt
appdirs==1.4.3
Django==1.10
django-filter==1.0.2
djangorestframework==3.6.2
httpie==0.9.9
Markdown==2.6.8
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
Pygments==2.2.0
pyparsing==2.2.0
reportlab==3.4.0
requests==2.14.2
six==1.10.0
